back/requirements.txt
1) download python 3.7 then execute in cmd :

pip install Django==2.2
django-cors-headers==3.3.0
django-countries==6.1.2
django-crispy-forms==1.9.0
django-phone-field==1.8.0
django-tinymce==3.0.2
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0

2)download node-js  then angular "npm install -g @angular/cli" then execute :

npm install ngx-bootstrap
alertifyjs@1.11.4
bootstrap
@auth0/angular-jwt
fontawesome
ng2-upload-file@1.3.0
ngx-gallery